+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 ### New Features :
 
 * Singleton Model. ([4bd47f9](https://github.com/NazioLT/NZCore/commit/4bd47f984792e7fdb75d74b28b530e73e3c865c3))
+* Child Monobehaviour.